Introduce Stream>>locale (a Locale for localising input/output)
This is to allow #displayOn: methods to perform localised output. This
requires a VM change to consolidate to a common instance variable because
of primitives that expect a fixed initial layout for the ReadStream and
WriteStream classes.

Added additional support to Locale to support further localised output, e.g
for COM Variants.

Add a Scintilla based view resource for the Transcript that is used when
the development system package is installed. This is to address the poor
performance of appending text to the multiline edit control.
Also Ensure that class comments and package comments are not ellipsised by
printString. For this purpose there is now fullPrintString.

fixup! Introduce (temporarily) Object>>displayOn:locale: