Include IP and source expression causing SUnit failure/error in test log
Some test methods may contain many assertions, and typically they don't
have a description that makes them easy to disambiguate, so it is useful
to include the source code of the expression that triggered either error
or assertion failure in the test method. If this is still ambiguous then
the actual IP is included, which case be used with the DebugInfoPlugin to
locate the point of failure.