Partition base system into namespaces; not including IDE yet
This is a fairly rough first cut. The IDE needs lot of work, especially
refactoring, but also auto-completion, class creation dialog, etc.
There is a problem with package pre-requisites not being recalculated
correctly after renaming a class, and after adding variables. This is
either due to a failure to flush cached binding references, or because
the package manager's map of variable names to packages has not been
updated correctly.